+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Fixed a potential NPE when collecting Uris in components
 
+### Changed
+- [#26] Small refactor of DomManager call
+
 ## [1.7.2]
 ### Changed
 - [#23] Updated and cleaned-up build.gradle file